+++ apps/decodex/src/agent/tracker_tool_bridge/tests/review_policy/review_policy_completion_gates/clean_checkpoint_gate/review_handoff_rejects_dirty_worktree_after_clean_checkpoint.rs
@@ -1,0 +1,53 @@
+use crate::agent::tracker_tool_bridge::tests::{
+	self, DynamicToolContentItem, DynamicToolHandler, FakeLocalRepoInspector,
+	FakePullRequestInspector, FakeTracker, ISSUE_REVIEW_HANDOFF_TOOL_NAME, TempDir,
+	TrackerToolBridge, review_policy,
+};
+
+#[test]
+fn review_handoff_rejects_dirty_worktree_after_clean_checkpoint() {
+	let temp_dir = TempDir::new().expect("tempdir should create");
+	let tracker = FakeTracker::new();
+	let issue = tests::sample_issue();
+	let workflow = tests::sample_workflow();
+	let inspector = FakePullRequestInspector::new(vec![Ok(tests::sample_pull_request())]);
+	let review_context = tests::sample_review_context_in(temp_dir.path());
+	let local_repo_inspector =
+		FakeLocalRepoInspector::new(vec![Ok(review_policy::sample_dirty_local_repo())]);
+	let bridge = TrackerToolBridge::with_review_handoff_for_test(
+		&tracker,
+		&issue,
+		&workflow,
+		review_context.clone(),
+		&inspector,
+		&local_repo_inspector,
+	);
+
+	tests::write_review_policy_checkpoint(
+		&bridge,
+		&issue,
+		&review_context,
+		"handoff",
+		"clean",
+		&tests::sample_local_repo().head_oid,
+		0,
+	);
+
+	let response = DynamicToolHandler::handle_call(
+		&bridge,
+		ISSUE_REVIEW_HANDOFF_TOOL_NAME,
+		serde_json::json!({
+			"pr_url": "https://github.com/hack-ink/decodex/pull/48",
+			"summary": "Ready for review."
+		}),
+	);
+
+	assert!(!response.success);
+	assert!(matches!(
+		response.content_items.as_slice(),
+		[DynamicToolContentItem::InputText{ text }]
+			if text.contains("requires a clean committed lane HEAD")
+				&& text.contains("record a fresh clean checkpoint")
+				&& text.contains("M apps/decodex/src/agent/tracker_tool_bridge/tools.rs")
+	));
+}
