+++ apps/decodex/src/orchestrator/tests/operator/status/running_lanes/ghost_lane/cleanup_blockers/live_evidence/ghost_lane_cleanup_status_blockers_reject_live_process_evidence.rs
@@ -1,0 +1,44 @@
+use crate::orchestrator::tests::operator::status::running_lanes::{
+	self, FakeTracker, StateStore, fs, orchestrator, process, state,
+};
+
+#[test]
+fn ghost_lane_cleanup_status_blockers_reject_live_process_evidence() {
+	let (_temp_dir, config, workflow) = running_lanes::temp_project_layout();
+	let state_store = StateStore::open_in_memory().expect("state store should open");
+	let tracker = FakeTracker::new(Vec::new());
+	let worktree_path = config.worktree_root().join("PUB-012");
+
+	fs::create_dir_all(&worktree_path).expect("worktree path should exist");
+
+	state_store
+		.record_run_attempt("run-12", "PUB-012", 1, "running")
+		.expect("run attempt should record");
+	state_store
+		.upsert_lease("pubfi", "PUB-012", "run-12", "In Progress")
+		.expect("lease should record");
+	state_store
+		.upsert_worktree(
+			"pubfi",
+			"PUB-012",
+			"x/pubfi-pub-012",
+			&worktree_path.display().to_string(),
+		)
+		.expect("worktree should record");
+
+	state::write_run_activity_marker_for_process(&worktree_path, "run-12", 1, process::id())
+		.expect("live process marker should write");
+
+	let blockers = orchestrator::ghost_lane_cleanup_status_blockers(
+		&tracker,
+		&config,
+		&workflow,
+		&state_store,
+		"PUB-012",
+		"run-12",
+	)
+	.expect("cleanup status blockers should load");
+
+	assert!(blockers.contains(&String::from("process_alive")));
+	assert!(blockers.contains(&String::from("retained_worktree_present")));
+}
